+++ src/lib.rs
@@ -507,6 +507,7 @@ mod error;
 mod header;
 pub mod svq1_blocktree;
 pub mod svq1_codebook;
+pub mod svq1_enc;
 pub mod svq1_helper_luts;
 pub mod svq1_mc;
 pub mod svq1_mean;